+++ src/main/java/com/powerup/propertymicroservice/infrastructure/entities/HouseEntity.java
@@ -34,7 +34,7 @@ public class HouseEntity {
     private Integer numberOfRooms;
 
     @Column(nullable = false)
-    private BigDecimal numberOfBathrooms;
+    private Integer numberOfBathrooms;
 
     @Column(nullable = false)
     private BigDecimal price;
